backend/requirements.txt
fastapi
uvicorn
numpy
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
transformers
requests
python-dotenv
python-multipart